requirements.txt
uvicorn~=0.20.0
fastapi~=0.90.0
starlette~=0.23.0
openai~=0.26.5
pyttsx3~=2.90
pydantic~=1.10.4
translators~=5.5.6
python-dotenv~=0.21.1